Speed up stack walks significantly
Adding a call site cache that should significantly reduce the cost of
retrieving the call site for an operator.

This cheats by using an internal class in the BCL to get just the
methodhandle and IL offset and uses that as a cache key for the full
blown, expensive stack frame.